refactor: 담당 도메인 코드 품질 개선 - 미사용 변수 및 FSD import 정리 (#270)
