+++ omechu-app/src/app/(auth)/login/email/page.tsx
@@ -9,14 +9,15 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useQueryClient } from "@tanstack/react-query";
 import { useForm, Controller } from "react-hook-form";
 
-import { ApiClientError, getAuthErrorMessage } from "@/entities/user";
-import { fetchProfile } from "@/entities/user/api/profileApi";
-import { useLoginMutation } from "@/entities/user/lib/hooks/useAuth";
 import {
+  ApiClientError,
+  getAuthErrorMessage,
+  fetchProfile,
+  useLoginMutation,
   loginSchema,
-  LoginFormValues,
-} from "@/entities/user/model/auth.schema";
-import { useAuthStore } from "@/entities/user/model/auth.store";
+  type LoginFormValues,
+  useAuthStore,
+} from "@/entities/user";
 import { Button, FormField, Input, Toast, useToast } from "@/shared";
 
 /**
